Uat (#32)
* Added channels partial (#25)

* Added message embed splitter for very long messages (#29)

* Bug/throttling rapid copy paste (#31)

* Added message embed splitter for very long messages

* Prevented bot attempting to delete the same trigger message twice

+++ src/messages/deleteTriggerMessage.ts
@@ -5,8 +5,11 @@ import * as Discord from "../discord";
 const deleteTriggerMessage = async (message: Message) => {
   try {
     if (message.channel.type !== "DM") {
-      Discord.addDeletedMessageId(message.id);
-      await message.delete();
+      if (!Discord.isBotDeletedMessage(message.id)) {
+        Discord.addDeletedMessageId(message.id);
+
+        await message.delete();
+      }
     }
   } catch (e) {
     Log.error(e);
